requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
google-genai
python-dotenv
requests
streamlit